refactor(audio): split audio/mod.rs into cpal + tests peers (#597)
Item 5 of the architecture-cleanup roadmap. Pure refactor — no
behaviour change. Pre-emptively unblocks Linux (#106) and Windows
(#107) audio impls by getting the cpal worker out of the module
front door.

Before:
- audio/mod.rs at 1900 lines holding the trait, the cpal worker,
  the handle structs, ring-buffer plumbing, AND every test.

After:
- audio/mod.rs at 533 lines: types, traits, MAX_BUFFER_FRAMES,
  shared ring helpers (drain_consumer, log_overflow_if_set), and
  the module map.
- audio/cpal.rs at 841 lines: CpalAudioCapture worker + Cmd enum +
  CpalMicSessionHandle + Session + cpal-specific tests
  (i16/u16 conversions, rms math, push_samples overflow).
- audio/tests.rs at 581 lines: cross-platform trait tests + the
  shared-helper tests, peer-loaded via `#[cfg(test)] mod tests;`
  so mod.rs stays at the trait + types budget.

audio/core_audio_tap.rs (macOS system-audio) was already a peer; this
PR puts the cpal backend on the same shape. When Linux PulseAudio (#106)
or Windows WASAPI loopback (#107) lands, those will be sibling files
under the same pattern (audio/pulse.rs, audio/wasapi.rs) rather than
doubling the size of mod.rs.

The strict <500-line target for mod.rs landed at 533 — the overage is
in the (well-documented) trait definitions themselves, which are the
natural shape. The architectural goal — getting the cpal worker out
of the module front door — is achieved.

Public API unchanged:
- `crate::audio::CpalAudioCapture` still resolves (re-exported
  from `cpal::` peer).
- `crate::audio::{AudioCapture, AudioSession, AudioSource,
  CaptureFormat, CapturedAudio, AudioDevice, AudioSourceListing,
  AudioSourceKind}` all unchanged.

Verified:
- cargo check --all-targets clean
- cargo clippy --all-targets -- -D warnings clean
- cargo clippy --lib --no-default-features -- -D warnings clean
  (the cross-platform path from CLAUDE.md)
- cargo test --lib: 412 pass, 0 fail (same count as before)
